Polymarket-intelligence/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
requests>=2.31.0
discord-webhook>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
web3>=6.0.0
eth-abi>=4.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
pydantic>=2.0.0
py-clob-client>=0.18.0
streamlit>=1.32.0
plotly>=5.18.0